pyproject.toml
[project]
name = "multi-agents-router-knowledge-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "langchain>=0.3.28",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "langsmith[claude-agent-sdk]>=0.4.37",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.39.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
